anonimizacion_imagenes/requirements.txt
#pip install -r requirements.txt
#easyocr==1.7.2
#opencv-python-headless==4.11.0.86
#torch==2.6.0
#scikit-image==0.24.0
#torchvision==0.21.0

blinker==1.9.0
certifi==2025.1.31
click==8.1.8

exceptiongroup==1.2.2
filelock==3.17.0
PyJWT==2.3.0
Flask==3.1.0
flask==3.1.0
fsspec==2025.2.0
imageio==2.37.0
importlib_metadata==8.6.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
lazy_loader==0.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.2.1
ninja==1.11.1.3
numpy==2.0.2
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
pulsar-client==3.6.0
pulsar-client[avro]==3.6.0
pyclipper==1.3.0.post6
pytest==8.3.4
python-bidi==0.6.6
PyYAML==6.0.2
scipy==1.13.1
shapely==2.0.7
sympy==1.13.1
tifffile==2024.8.30
tomli==2.2.1
typing_extensions==4.12.2
Werkzeug==3.1.3
zipp==3.21.0
aiopulsar-py==0.1.0a0
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-core==2.4.2
google-cloud-storage==3.0.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.68.0
flask_sqlalchemy==3.0.0
psycopg2-binary==2.9.10
marshmallow_sqlalchemy==0.26.1